-create functons in cubit to filter doctors -create 2 states
